+++ src/Infrastructure/Repository/ChatSessionRepository.php
@@ -6,6 +6,7 @@ use Doctrine\ORM\EntityManager;
 use Proximum\Vimeet\Domain\Model\ChatMessage;
 use Proximum\Vimeet\Domain\Model\ChatSession;
 use Proximum\Vimeet\Domain\Model\Event;
+use Proximum\Vimeet\Domain\Model\Sheet;
 use Proximum\Vimeet\Domain\Model\User;
 use Proximum\Vimeet\Domain\Repository\ChatSessionRepositoryInterface;
 
@@ -136,6 +137,21 @@ class ChatSessionRepository implements ChatSessionRepositoryInterface
         return !empty($queryBuilder->getQuery()->getArrayResult());
     }
 
+    public function hasAStartedVisio (Event $event, User $user): bool
+    {
+        $queryBuilder = $this->entityManager->createQueryBuilder()
+            ->select('session.id')
+            ->from(ChatSession::class, 'session')
+            ->andWhere('session.visioStartedAt IS NOT NULL')
+            ->andWhere('session.fromUser = :user OR session.toUser = :user')
+            ->setParameter('user', $user)
+            ->andWhere('session.event = :event')
+            ->setParameter('event', $event)
+            ->setMaxResults(1);
+
+        return null !== $queryBuilder->getQuery()->getOneOrNullResult();
+    }
+
     public function countCallVisioByEvent(Event $event): int
     {
         $queryBuilder = $this->entityManager->createQueryBuilder()
@@ -147,6 +163,7 @@ class ChatSessionRepository implements ChatSessionRepositoryInterface
             ->setMaxResults(1);
 
         return $queryBuilder->getQuery()->getSingleScalarResult();
+
     }
 
 }